consolidate spruce spa dependencies and peer dependencies install;
